backend/requirements.txt
annotated-doc==0.0.4
moorcheh-sdk>=1.3.0
python-dotenv>=1.0.0
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==1.3.5
langchain-core==1.2.19
langchain-openai==1.1.11
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
openai==2.28.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
xxhash==3.6.0
zstandard==0.25.0
